intel_poller: Removed explicit jsonb spec
